+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Mark events as "tracked" to prevent double tracking ([PR #4395](https://github.com/alphagov/govuk_publishing_components/pull/4395))
 * Search with autocomplete: track non-empty suggestions ([PR #4400](https://github.com/alphagov/govuk_publishing_components/pull/4400))
 * Add heading option to metadata component ([PR #4383](https://github.com/alphagov/govuk_publishing_components/pull/4383))
+* Track autocomplete acceptance in session storage ([PR #4381](https://github.com/alphagov/govuk_publishing_components/pull/4381))
 
 ## 45.2.0